pyproject.toml
[tool.poetry]
name = "storage-area-based-on-fastapi"
version = "0.1.0"
description = ""
authors = ["AniutaP <musenok.ko@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
python-multipart = "^0.0.12"
aiosqlite = "^0.20.0"
sqlalchemy = "^2.0.35"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
black = "^24.10.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.0"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
dacite = "^1.8.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
